fix set execution in nanosecond resolution

+++ readAndExecute.go
@@ -72,7 +72,7 @@ output=%s
 func readAndExecute(received *receivedStruct, config *configurationStruct) *answer {
 	var result answer
 	// first set the start time
-	result.startTime = float64(time.Now().UnixNano() / int64(time.Second))
+	result.startTime = float64(time.Now().UnixNano()) / float64(time.Second)
 	result.source = "Mod-Gearman Worker @ " + config.identifier
 	result.active = "active"
 	result.resultQueue = received.resultQueue
@@ -110,7 +110,7 @@ func readAndExecute(received *receivedStruct, config *configurationStruct) *answ
 	}
 
 	// last set the finish time
-	result.finishTime = float64(time.Now().UnixNano() / int64(time.Second))
+	result.finishTime = float64(time.Now().UnixNano()) / float64(time.Second)
 
 	return &result
 }
